refactor(toaster): fixed toaster file to align with fast-refresh
