+++ apps/client/src/chakra/ui/toaster/toaster-store.ts
@@ -1,0 +1,6 @@
+import { createToaster } from '@chakra-ui/react';
+
+export const toaster = createToaster({
+  placement: 'top',
+  pauseOnPageIdle: true,
+});
