+++ apps/client/src/hooks/query/useVerifyCode.ts
@@ -2,7 +2,7 @@ import { useMutation } from '@tanstack/react-query';
 import { API } from '../../axios';
 import { VerifyPhoneNumberDto } from '@komuna/types';
 import { type AxiosError } from 'axios';
-import { toaster } from '../../chakra/ui/toaster';
+import { toaster } from '../../chakra/ui/toaster/toaster-store';
 import { t } from 'i18next';
 
 interface VerifyCodeResponse {
